requirements.txt
fastapi>=0.115.0
uvicorn>=0.30.0
pydantic>=2.8.0
python-dotenv>=1.0.1
boto3>=1.35.0
neo4j>=5.28.0
neo4j-graphrag>=1.14.0
openai>=1.40.0
langchain
langchain-openai
deepagents
pytest>=8.3.0
